Rename components to Pascal case
