+++ app/layout.tsx
@@ -1,4 +1,4 @@
-import NavBar from '@/components/navbar';
+import NavBar from '@/components/Navbar';
 import './globals.css';
 import type { Metadata } from 'next';
