+++ components/GalleryWindow.tsx
