+++ components/LaptopAnimation.tsx
